refactor(visual): extract fluid simulator scene into composable
Move the WebGPU water/sky scene setup out of the page and into
useFluidSimulator, leaving the page as a bill of materials.
